requirements.txt
datasets==2.16.1
joblib==1.3.2
numpy==1.26.3
pandas==2.1.4
scikit_learn==1.3.2
tensorflow==2.15.0
tensorflow_intel==2.15.0
umap_learn==0.5.5
